Attached CSS to all pages, Bottom Nav, and other edits
CSS is now attached to each page.
Bottom Nav is now in footer. I also made changes to the email, copyright year, modification date, and W3C approval badge to the footer of each page.
Edited the Birthday subscription feature so it accepts a date.
Replaced the header with an image.
Edited navs on each page so it was not in an unordered list.
Changed the display of the About us page to a center view. I liked the center over left and right.
Added 100% width to the regions class to fix the centering issue we were having.
For the recipe tables I edited the colspan to 3, because when I made the adjustments to altering left and right alignment it looked neater. Also to alter left and right I moved the image in the td table up to the top in the Main recipe.
In the Norway/Ghana links I added a div a:link features and added a background color in the div.gallery.
Altered the text sizes in css.
Centered the alignment for the Reference page as well.